Extracted each section from INPO
Each file includes both the definition and the example. Before sending Bard for vectorization, we could consider extracting those examples and labeling those examples so they could be used as official labels for training SVM purposes. Additionally, as we have some labels that are the minority in the annual report, here we can have some official examples for those as well, helping us to potentially synthesize more examples in the future.